@@ -3,28 +3,37 @@
   "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "leaflet": "^1.0.3",
-    "material-ui": "^0.17.4",
-    "moment": "^2.18.1",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
-    "react-leaflet": "^1.1.6",
-    "react-redux": "^5.0.4",
-    "react-router-dom": "^4.1.1",
-    "react-tap-event-plugin": "^2.0.1",
-    "redux": "^3.6.0",
-    "redux-actions": "^2.0.2",
-    "redux-thunk": "^2.2.0",
-    "rrule": "^2.2.0",
-    "superagent": "^3.5.2"
+    "@material-ui/core": "^3.9.3",
+    "@material-ui/icons": "^3.0.2",
+    "leaflet": "^1.4.0",
+    "moment": "^2.24.0",
+    "prop-types": "^15.7.2",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "react-leaflet": "^2.2.1",
+    "react-redux": "^6.0.1",
+    "react-router-dom": "^5.0.0",
+    "react-tap-event-plugin": "^3.0.3",
+    "react-virtualized": "^9.21.0",
+    "redux": "^4.0.1",
+    "redux-actions": "^2.6.5",
+    "redux-thunk": "^2.3.0",
+    "rrule": "^2.6.0",
+    "superagent": "^5.0.2"
   },
   "devDependencies": {
-    "react-scripts": "0.9.5"
+    "react-scripts": "2.1.8"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
-  }
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
